nvmecontrol: Apply cast
The proper fix also casts to uintmax_t.

Reported by:	imp
Fixes:		4c9ab06813b5

+++ sbin/nvmecontrol/modules/samsung/samsung.c
@@ -146,8 +146,8 @@ print_samsung_extended_smart(const struct nvme_controller_data *cdata __unused, 
 	print_temp(le16dec(&temp->ct));
 	printf("   Capacitor Health                         : %u\n", le16dec(&temp->ch));
 	printf("   Reserved Erase Block Count               : %u\n", le32dec(&temp->luurb));
-	printf("   Read Reclaim Count                       : %ju\n", le64dec(&temp->rrc));
-	printf("   Lifetime Uncorrectable ECC Count         : %ju\n", le64dec(&temp->lueccc));
+	printf("   Read Reclaim Count                       : %ju\n", (uintmax_t) le64dec(&temp->rrc));
+	printf("   Lifetime Uncorrectable ECC Count         : %ju\n", (uintmax_t) le64dec(&temp->lueccc));
 	printf("   Reallocated Block Count                  : %u\n", le32dec(&temp->lurb));
 	printf("   Power on Hours                           : %s\n",
 	    uint128_to_str(to128(temp->poh), cbuf, sizeof(cbuf)));
